Minor bugfixes in web client
Change the starting sequence number when sending a link to 1 to avoid
issues with TOFU and clients added through the discovery process.
When trusting a device through the discovery, the device is added to the
TOFU with a starting seq number of 0, so a message received with a 0 seq
number from that key will be discarded. Using 1 instead fixes this.